+++ packages/icons/src/index.js
@@ -43,7 +43,6 @@ export { default as cloudUpload } from './library/cloud-upload';
 export { default as cloud } from './library/cloud';
 export { default as code } from './library/code';
 export { default as cog } from './library/cog';
-export { default as cogAlt } from './library/cog-alt';
 export { default as color } from './library/color';
 export { default as column } from './library/column';
 export { default as columns } from './library/columns';